Rearrange & redesign some chat stuffs
