+++ src/IronworksTranslator/Views/Pages/DeveloperPage.xaml
@@ -22,6 +22,16 @@
                 <Border Width="8"/>
                 <ui:Button Content="Remove one and add chat" Command="{Binding ViewModel.DietChatCommand}"/>
             </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <ui:TextBlock Margin="8,0,0,0" Text="Chat drag: " VerticalAlignment="Center"/>
+                <ui:Button Margin="16,0,0,0" Content="Enable" Click="EnableChatWindowDragButton_Click"/>
+                <ui:Button Margin="16,0,0,0" Content="Disable" Click="DisableChatWindowDragButton_Click"/>
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <ui:TextBlock Margin="8,0,0,0" Text="Chat window grip: " VerticalAlignment="Center"/>
+                <ui:Button Margin="16,0,0,0" Content="Enable" Click="EnableChatWindowGripButton_Click"/>
+                <ui:Button Margin="16,0,0,0" Content="Disable" Click="DisableChatWindowGripButton_Click"/>
+            </StackPanel>
         </StackPanel>
     </Grid>
 </Page>
