+++ src/IronworksTranslator/Views/Pages/DeveloperPage.xaml.cs
@@ -1,4 +1,5 @@
 ﻿using IronworksTranslator.ViewModels.Pages;
+using IronworksTranslator.Views.Windows;
 using Wpf.Ui.Controls;
 
 namespace IronworksTranslator.Views.Pages
@@ -16,6 +17,30 @@ namespace IronworksTranslator.Views.Pages
             DataContext = this;
 
             InitializeComponent();
+        }
+
+        private void DisableChatWindowGripButton_Click(object sender, RoutedEventArgs e)
+        {
+            var chatWindow = App.GetService<ChatWindow>();
+            chatWindow.ResizeMode = ResizeMode.NoResize;
+        }
+
+        private void EnableChatWindowGripButton_Click(object sender, RoutedEventArgs e)
+        {
+            var chatWindow = App.GetService<ChatWindow>();
+            chatWindow.ResizeMode = ResizeMode.CanResizeWithGrip;
+        }
+
+        private void EnableChatWindowDragButton_Click(object sender, RoutedEventArgs e)
+        {
+            var chatWindow = App.GetService<ChatWindow>();
+            chatWindow.ViewModel.IsDraggable = true;
+        }
+
+        private void DisableChatWindowDragButton_Click(object sender, RoutedEventArgs e)
+        {
+            var chatWindow = App.GetService<ChatWindow>();
+            chatWindow.ViewModel.IsDraggable = false;
         }
     }
 }
